document-converter-service/requirements.txt
uvicorn[standard]
fastapi
loguru==0.6.0
requests==2.28.2
aspose-words==23.1.0
bs4==0.0.1